requirements.txt
argparse
colorama
pbr
GitPython==0.3.2.RC1
giturlparse.py
pyyaml
jinja2
requests
pygerrit
docker-py